Another set of fixes
